requirements.txt
customtkinter>=5.2.0
svgwrite>=1.4.3
reportlab>=4.0.0
python-pptx>=0.6.21
svglib>=1.5.1
Pillow>=10.0.0
google-genai>=0.1.0
numpy>=1.24.0
python-magic-bin; platform_system == 'Windows'
python-magic; platform_system != 'Windows'
